# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Api of WxMall",
        "title": "Golang WxMall API",
        "contact": {},
        "license": {
            "name": "Apache 2.0"
        },
        "version": "1.0"
    },
    "basePath": "127.0.0.1:8000",
    "paths": {
...
        "/api/v1/product/index": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "商品列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "brand品牌",
                        "name": "brand",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "series系列",
                        "name": "series",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "page页数",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/product/info/:id": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "根据id获取商品信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/product/search": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "搜索",
                "parameters": [
                    {
                        "type": "string",
                        "description": "根据标签搜索",
                        "name": "tag",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "根据商品名搜索",
                        "name": "title",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/user/order": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "获取所有订单信息",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "新建订单",
                "parameters": [
                    {
                        "description": "购物车IDs",
                        "name": "cartIds",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "地址ID",
                        "name": "addressId",
                        "in": "body",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/user/order/:id/pay": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "支付订单",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "要支付的订单id",
                        "name": "oid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/user/order/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "查看订单包含的商品",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "要查询的订单id",
                        "name": "oid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "summary": "销毁订单",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "要移除的id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "app.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}